linux: move libnl code into separate file
Move all the code using libnl functionality into a separate file to ease
modifications.

No functional change.
